src/main/java/com/example/weatherapp/controller/WeatherController.java
package com.example.weatherapp.controller;

import com.example.weatherapp.service.WeatherService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/weather")
public class WeatherController {

    private final WeatherService weatherService;

    @Autowired
    public WeatherController(WeatherService weatherService) {
        this.weatherService = weatherService;
    }

    @GetMapping("/{zipCode}")
    public ResponseEntity<String> getWeatherByZipCode(@PathVariable("zipCode") String zipCode) {
        try {
            String weatherData = weatherService.getWeatherDetails(zipCode);
            return ResponseEntity.ok(weatherData);
        } catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                                 .body("Failed to fetch weather data");
        }
    }

    // Add more endpoints as needed for forecast, historical data, etc.
}
